+++ src/main/java/tudelft/ti2806/pl3/data/filter/ViewFilter.java
@@ -1,6 +1,6 @@
 package tudelft.ti2806.pl3.data.filter;
 
-import tudelft.ti2806.pl3.data.graph.Node;
+import tudelft.ti2806.pl3.data.graph.node.Node;
 
 import java.util.List;
