+++ src/main/java/tudelft/ti2806/pl3/data/graph/GraphData.java
@@ -1,6 +1,7 @@
 package tudelft.ti2806.pl3.data.graph;
 
 import tudelft.ti2806.pl3.data.Genome;
+import tudelft.ti2806.pl3.data.graph.node.Node;
 
 import java.util.List;
