+++ src/main/java/tudelft/ti2806/pl3/data/graph/node/Node.java
@@ -1,4 +1,4 @@
-package tudelft.ti2806.pl3.data.graph;
+package tudelft.ti2806.pl3.data.graph.node;
 
 import tudelft.ti2806.pl3.data.Genome;
