+++ src/main/java/tudelft/ti2806/pl3/data/graph/node/SingleNode.java
@@ -1,4 +1,4 @@
-package tudelft.ti2806.pl3.data.graph;
+package tudelft.ti2806.pl3.data.graph.node;
 
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import tudelft.ti2806.pl3.data.Genome;
