+++ src/main/java/tudelft/ti2806/pl3/visualization/GraphController.java
@@ -1,7 +1,7 @@
 package tudelft.ti2806.pl3.visualization;
 
 import tudelft.ti2806.pl3.data.filter.Filter;
-import tudelft.ti2806.pl3.data.graph.Node;
+import tudelft.ti2806.pl3.data.graph.node.Node;
 
 import java.awt.Component;
 import java.util.ArrayList;
