+++ src/test/java/tudelft/ti2806/pl3/data/graph/GraphParserTest.java
@@ -6,6 +6,8 @@ import org.junit.Test;
 
 import tudelft.ti2806.pl3.data.BasePair;
 import tudelft.ti2806.pl3.data.Genome;
+import tudelft.ti2806.pl3.data.graph.node.Node;
+import tudelft.ti2806.pl3.data.graph.node.SingleNode;
 
 import java.io.*;
 import java.util.HashMap;
